#61 As an admin, I want to be able to schedule a monthly newsletter to be sent automatically, detailing the available classes for the coming month.
Definition of Done:

Auto-send Monthly Newsletter

Functionality should be added, on top of the current newsletter system, to automatically send a monthly newsletter.

On the newsletter page of the admin dashboard, the admin should be able to set the day of the month on which the newsletter is sent. For example, admin can choose newsletter to be sent on the first of every month or the fifth of every month, etc.

On the newsletter page, the admin should be able to preview the format and content of the currently scheduled monthly newsletter.

The course data that makes up the monthly newsletter will be pulled from the database. The newsletter will select all courses that are available for the month ahead, starting with the day selected as the send date.

These courses will be presented in an easy to read format, displaying the course photo, description, and date.

Saving Newsletter Templates to Database

Our current implementation sends out dummy templates to the subscriber list. Now that the newsletter functionality is in place, the admin should have realistic templates that can be filled with relevant information and sent to the subscriber list.

This sprint, I will focus on implementing two of these templates. One template for the monthly automated newsletter and one template to highlight just one specific upcoming event.

The format of these templates will be based off of the clients current newsletter system.

These templates will be saved to the database and be accessible both on the newsletter admin page and on the template editor page.

On the newsletter admin page, the finished newsletters will be shown on the currently implemented drop down menu. Once selected, the admin can send the newsletter to the subscriber list just as before.

On the newsletter editor page, the templates will be listed on the side bar. The admin can then select a template from the side bar and edit it’s contents, such as text, pictures, header, etc. Once finished, the admin can press the “save” button to save the current template to the database and mark it as completed.

Overall Goal

Allow the admin to send relevant information to the subscriber list through well designed templates.

Make the editing experience simple for the admin so that they only need to fill a box of text or swap out a photo.

Ensure that customers on the subscriber list are made aware of all upcoming courses being taught in the coming month.

Acceptance Criteria

Newsletter is automatically sent at the scheduled interval.

Templates are accurately saved and retrieved from the database.

Newsletter contains relevant and accurate information about courses, pulled from the database.

Practical Summary:
Implemented automated monthly newsletter sending.
Modified email builder to allow saving templates to database.
Modified email builder to fetch templates from database and display in side panel.
Modified newsletter admin page to fetch templates from database and display in dropdown menu.
Added email template database model and appropriate api routes.
Added client side utils for interactive with api routes.